update in init for mpcs

+++ flexibility_quantification/modules/baseline_mpc.py
@@ -22,14 +22,19 @@ class FlexibilityBaselineMPC(mpc_full.MPC):
 
     def __init__(self, config, agent):
         super().__init__(config, agent)
-        # clear the casadi simulator result at the first time step
-        self.res_file_flex = self.config.optimization_backend['results_file'].replace('mpc', 'mpc_sim')
-        # initialize the flex_model for integration
-        self.flex_model = type(self.model)(dt=self.config.casadi_sim_time_step)
-        try:
-            os.remove(self.res_file_flex)
-        except:
-            pass
+        # initialize flex_results with None
+        self.flex_results = None
+        # set up necessary components if simulation is enabled
+        if self.config.casadi_sim_time_step > 0:
+            # generate a separate flex_model for integration to ensure the model used in MPC optimization remains unaffected
+            self.flex_model = type(self.model)(dt=self.config.casadi_sim_time_step)
+            # generate the filename for the simulation results
+            self.res_file_flex = self.config.optimization_backend['results_file'].replace('mpc', 'mpc_sim')
+            # clear the casadi simulator result at the first time step if already exists
+            try:
+                os.remove(self.res_file_flex)
+            except:
+                pass
 
 
     def pre_computation_hook(self):
@@ -43,7 +48,7 @@ class FlexibilityBaselineMPC(mpc_full.MPC):
 
     def set_output(self, solution):
         """Takes the solution from optimization backend and sends it to AgentVariables."""
-        # Output must be defined in the conig as "type"="pd.Series"
+        # Output must be defined in the config as "type"="pd.Series"
         if not self.config.set_outputs:
             return
         self.logger.info("Sending optimal output values to data_broker.")
@@ -52,7 +57,7 @@ class FlexibilityBaselineMPC(mpc_full.MPC):
         self.sim_flex_model(solution)
 
         df = solution.df
-        if hasattr(self, "flex_results"):
+        if self.flex_results is not None:
             for output in self.var_ref.outputs:
                 if output not in [self.config.power_variable_name, self.config.storage_variable_name]:
                     series = df.variable[output]
@@ -106,7 +111,7 @@ class FlexibilityBaselineMPC(mpc_full.MPC):
         else:
             store_results_df.to_csv(self.res_file_flex, mode='a', header=False)
 
-        # set the flex results format same as mpc result while updating Agengvariable
+        # set the flex results format same as mpc result while updating Agentvariable
         self.flex_results.index = self.flex_results.index.get_level_values(1)
 
     def _initialize_flex_results(self, n_simulation_steps, horizon_length, sim_time_step, result_df):
@@ -124,7 +129,7 @@ class FlexibilityBaselineMPC(mpc_full.MPC):
             names=['time_step', 'time'])
         # merge indexes
         new_index = index_coll.union(index_full_sample).sort_values()
-        # initialize the flex results
+        # initialize the flex results with correct dimension
         self.flex_results = pd.DataFrame(np.nan, index=new_index,
                                          columns=self.var_ref.outputs)
 
@@ -207,14 +212,19 @@ class FlexibilityBaselineMINLPMPC(minlp_mpc.MINLPMPC):
 
     def __init__(self, config, agent):
         super().__init__(config, agent)
-        # clear the casadi simulator result at the first time step
-        self.res_file_flex = self.config.optimization_backend['results_file'].replace('mpc', 'mpc_sim')
-        # initialize the flex_model for integration
-        self.flex_model = type(self.model)(dt=self.config.casadi_sim_time_step)
-        try:
-            os.remove(self.res_file_flex)
-        except:
-            pass
+        # initialize flex_results with None
+        self.flex_results = None
+        # set up necessary components if simulation is enabled
+        if self.config.casadi_sim_time_step > 0:
+            # generate a separate flex_model for integration to ensure the model used in MPC optimization remains unaffected
+            self.flex_model = type(self.model)(dt=self.config.casadi_sim_time_step)
+            # generate the filename for the simulation results
+            self.res_file_flex = self.config.optimization_backend['results_file'].replace('mpc', 'mpc_sim')
+            # clear the casadi simulator result at the first time step if already exists
+            try:
+                os.remove(self.res_file_flex)
+            except:
+                pass
 
     def pre_computation_hook(self):
         if self.get("in_provision").value:
@@ -229,7 +239,7 @@ class FlexibilityBaselineMINLPMPC(minlp_mpc.MINLPMPC):
 
     def set_output(self, solution):
         """Takes the solution from optimization backend and sends it to AgentVariables."""
-        # Output must be defined in the conig as "type"="pd.Series"
+        # Output must be defined in the config as "type"="pd.Series"
         if not self.config.set_outputs:
             return
         self.logger.info("Sending optimal output values to data_broker.")
@@ -238,7 +248,7 @@ class FlexibilityBaselineMINLPMPC(minlp_mpc.MINLPMPC):
         self.sim_flex_model(solution)
 
         df = solution.df
-        if hasattr(self, "flex_results"):
+        if self.flex_results is not None:
             for output in self.var_ref.outputs:
                 if output not in [self.config.power_variable_name, self.config.storage_variable_name]:
                     series = df.variable[output]
@@ -292,7 +302,7 @@ class FlexibilityBaselineMINLPMPC(minlp_mpc.MINLPMPC):
         else:
             store_results_df.to_csv(self.res_file_flex, mode='a', header=False)
 
-        # set the flex results format same as mpc result while updating Agengvariable
+        # set the flex results format same as mpc result while updating Agentvariable
         self.flex_results.index = self.flex_results.index.get_level_values(1)
 
     def _initialize_flex_results(self, n_simulation_steps, horizon_length, sim_time_step, result_df):
@@ -310,7 +320,7 @@ class FlexibilityBaselineMINLPMPC(minlp_mpc.MINLPMPC):
             names=['time_step', 'time'])
         # merge indexes
         new_index = index_coll.union(index_full_sample).sort_values()
-        # initialize the flex results
+        # initialize the flex results with correct dimension
         self.flex_results = pd.DataFrame(np.nan, index=new_index,
                                          columns=self.var_ref.outputs)
